Use `current` variable instead of using a partial
It used to maintain a partial named `current`. Now it uses `current` on
ractive's data object to point to a file partial.

Ractive has a feature (that isn't very well documented) where if you have
{{>current}} in your template, and there is not a `current` partial, then it
will look for `current` in data. If found, it will act as a pointer to the
partial that you actually want.

So if you have 2 partials, file1.md="lol" and file2.md="wat", and your template
is "{{>current}}", and you do `ractive.set('current', 'file1.md')` then you'll
see "lol". If you then do `ractive.set('current', 'file2.md')`, then you'll see
"wat".

This feature works quite nicely in this circumstance because we were already
keeping the `current` data value up to date. Ractive 0.7.3 has a bug where
having a partial and a data attribute with the same name basically did not
work. This is a workaround, but it also reduces code.